+++ libs/event/test.c
@@ -85,7 +85,7 @@ static void test_am_event_queue(const int capacity, const int rdwr_num) {
     for (int i = 1; i < rdwr_num; ++i) {
         bool rc = am_event_push_back(&q, &events[i]);
         AM_ASSERT(rc);
-        AM_ASSERT(am_event_queue_get_nbusy(&q) == i);
+        AM_ASSERT(am_event_queue_get_nbusy_unsafe(&q) == i);
         AM_ASSERT(!am_event_queue_is_empty(&q));
     }
 
@@ -97,8 +97,8 @@ static void test_am_event_queue(const int capacity, const int rdwr_num) {
     for (int i = 1; i <= rdwr_num; ++i) {
         bool rc = am_event_push_front(&q, &events[i]);
         AM_ASSERT(rc);
-        AM_ASSERT(am_event_queue_get_nbusy(&q) > 0);
-        AM_ASSERT(am_event_queue_get_nbusy(&q) == i);
+        AM_ASSERT(am_event_queue_get_nbusy_unsafe(&q) > 0);
+        AM_ASSERT(am_event_queue_get_nbusy_unsafe(&q) == i);
         AM_ASSERT(!am_event_queue_is_empty(&q));
     }
 
@@ -107,7 +107,7 @@ static void test_am_event_queue(const int capacity, const int rdwr_num) {
         AM_ASSERT(&events[i] == event);
     }
 
-    AM_ASSERT(am_event_queue_get_nbusy(&q) == 0);
+    AM_ASSERT(am_event_queue_get_nbusy_unsafe(&q) == 0);
     AM_ASSERT(am_event_queue_is_empty(&q));
 }